git-svn-id: http://monkey-stuff.googlecode.com/svn/trunk@155 aaeee498-0047-0410-bd38-07bdd6096719

+++ jnpr/nsm/nsmdiffemail.py
@@ -10,9 +10,10 @@ receivers = ['mlukaszuk@company.gov']
 
 MAINDIR = "/home/case/store/company/nsmdiff/.tosend"
 
-for filename in os.listdir(MAINDIR):
-  message = open(MAINDIR+"/"+filename,'r').read()
- 
-  smtpObj = smtplib.SMTP('mail.company.gov')
-  smtpObj.sendmail(sender, receivers, message)
-  os.unlink(MAINDIR+"/"+filename)  
+if os.path.exists(MAINDIR):
+  for filename in os.listdir(MAINDIR):
+    message = open(MAINDIR+"/"+filename,'r').read()
+   
+    smtpObj = smtplib.SMTP('mail.company.gov')
+    smtpObj.sendmail(sender, receivers, message)
+    os.unlink(MAINDIR+"/"+filename)
